==> src/pipeline/02_bronze_ingest_raw.ipynb <==
## 2. Descomplicando a Jornada dos Dados - Do Início à Camada Bronze

Já fizemos o processo de geração dos **perfis de cliente**, **terminal** e das **transações**. Além disso, colocamos em prática a **ingestão dos dados brutos** para a **camada raw**.

É importante lembrar que, por enquanto, estamos realizando esse processo com **dados simulados**. No entanto, se estivéssemos coletando esses dados de uma **API** e realizando algum tipo de tratamento, correríamos o risco de **interromper a pipeline** e **perder informações** caso a API sofresse alguma modificação.

A próxima etapa é pegar esses **dados brutos** da **camada raw** (que estão em formato CSV) e depositá-los na **camada bronze**. Nesta fase, não faremos muitas modificações: apenas a **troca do formato para Delta** e a **reorganização do cabeçalho das tabelas** para a posição correta.

custormer_df = spark.read.csv('/Volumes/workspace/fraud_detection/raw/customer/customer.csv', header=True)
terminal_df = spark.read.csv('/Volumes/workspace/fraud_detection/raw/terminal/terminal.csv', header=True)
transaction_df = spark.read.csv('/Volumes/workspace/fraud_detection/raw/transactions/transactions.csv', header=True)

# Tabelas de perfis
custormer_df.write.format("delta") \
    .mode("overwrite") \
    .saveAsTable("fraud_detection.customer_profiles_bronze")

terminal_df.write.format("delta") \
    .mode("overwrite") \
    .saveAsTable("fraud_detection.terminal_profiles_bronze")

# Tabelas de trasações
transaction_df.write.format("delta") \
    .mode("overwrite") \
    .saveAsTable("fraud_detection.transaction_bronze")

Como podemos ver nas **queries abaixo**, nossas tabelas já estão estruturadas e prontas para dar continuidade ao **tratamento dos dados** na **camada silver**.

%sql
SELECT
  *
FROM
  workspace.fraud_detection.customer_profiles_bronze
LIMIT 3

%sql
SELECT
  *
FROM
  workspace.fraud_detection.terminal_profiles_bronze
LIMIT 3

%sql
SELECT
  *
FROM
  workspace.fraud_detection.transaction_bronze
ORDER BY `TRANSACTION_ID` ASC
LIMIT 3
